requirements.txt
Django==3.0.5
djangorestframework==3.12.4
python-dotenv==0.17.0
